@@ -4,22 +4,26 @@
     <modelVersion>4.0.0</modelVersion>
     <groupId>org.panteleyev</groupId>
     <artifactId>password-manager</artifactId>
-    <version>20.1.1</version>
+    <version>20.2.0</version>
     <packaging>jar</packaging>
 
     <properties>
-        <app.main.class>org.panteleyev.pwdmanager.PasswordManagerApplication</app.main.class>
+        <appModule>password.manager</appModule>
+        <appMainClass>org.panteleyev.pwdmanager.PasswordManagerApplication</appMainClass>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <maven.compiler.source>13</maven.compiler.source>
-        <maven.compiler.target>13</maven.compiler.target>
+        <maven.compiler.source>14</maven.compiler.source>
+        <maven.compiler.target>14</maven.compiler.target>
         <commons-version>20.1.1</commons-version>
-        <javafx-version>13.0.1</javafx-version>
+        <javafx-version>14</javafx-version>
         <testng-version>6.14.3</testng-version>
         <!-- Plugins -->
         <mavenToolchainPluginVersion>3.0.0</mavenToolchainPluginVersion>
         <mavenCompilerPluginVersion>3.8.1</mavenCompilerPluginVersion>
         <mavenJarPluginVersion>3.2.0</mavenJarPluginVersion>
+        <mavenSurefirePluginVersion>3.0.0-M4</mavenSurefirePluginVersion>
         <mavenDependencyPluginVersion>3.1.1</mavenDependencyPluginVersion>
+        <javaFxPluginVersion>0.0.4</javaFxPluginVersion>
+        <execMavenPluginVersion>1.6.0</execMavenPluginVersion>
     </properties>
 
     <dependencyManagement>
